requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
pyyaml>=6.0.0
pydantic>=2.0.0
numpy>=2.0.0